feat(engine): change pixel diff_count from u32 to u64
- Update MismatchDetail variants (Pixel, SsimFallback) to use u64
- Change compare_pixels and count_mismatched_pixels return types to u64
- Fix pixel calculation to avoid overflow on large images
- Add validation to prevent relative path navigation in test names
- Improve entry handling in temp_cases to avoid unwrap
- Optimize path separator normalization to avoid allocations
- Add comprehensive tests for new behavior

+++ gleon-core/src/report.rs
@@ -217,7 +217,7 @@ impl<'a> Serialize for HtmlFailureView<'a> {
                     },
                 )?;
                 state.serialize_field("diff_path", &None::<String>)?;
-                state.serialize_field("diff_count", &None::<u32>)?;
+                state.serialize_field("diff_count", &None::<u64>)?;
                 state.serialize_field(
                     "actual_size",
                     &format!("{}x{}", actual_size.0, actual_size.1),
